requirements.txt
dynmen>=0.1.5
pydub>=0.24.0
pyyaml>=5.3
traitlets>=5.0.0
xdg>=5.0.0
